chore: refactor internal package structure
remove internal/dash0, move everything in internal/dash0 one level up
