+++ config/webhook/manifests.yaml
@@ -1,5 +1,5 @@
 # Note: The kustomize configs are not used in production, but they are currently still used when setting up the unit
-# test environment, like in internal/dash0/webhooks/webhook_suite_test.go -> &envtest.Environment#WebhookInstallOptions.
+# test environment, like in internal/webhooks/webhook_suite_test.go -> &envtest.Environment#WebhookInstallOptions.
 # Thus, they need to be kept in sync with the webhook resources defined in
 # helm-chart/dash0-operator/templates/operator/deployment-and-webhooks.yaml.
 ---
