+++ internal/predelete/operator_pre_delete_handler.go
