+++ internal/startup/auto_operator_configuration_handler_test.go
