+++ internal/workloads/workload_modifier_test.go
@@ -12,7 +12,7 @@ import (
 
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
-	"github.com/dash0hq/dash0-operator/internal/dash0/util"
+	"github.com/dash0hq/dash0-operator/internal/util"
 
 	. "github.com/dash0hq/dash0-operator/test/util"
 )
