+++ test/e2e/e2e_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/google/uuid"
 
 	dash0v1alpha1 "github.com/dash0hq/dash0-operator/api/dash0monitoring/v1alpha1"
-	"github.com/dash0hq/dash0-operator/internal/dash0/startup"
+	"github.com/dash0hq/dash0-operator/internal/startup"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
