+++ test/util/matchers.go
@@ -9,7 +9,7 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 
-	"github.com/dash0hq/dash0-operator/internal/dash0/util"
+	"github.com/dash0hq/dash0-operator/internal/util"
 
 	"github.com/onsi/gomega"
 	"github.com/onsi/gomega/format"
