+++ test/util/verification.go
@@ -16,7 +16,7 @@ import (
 
 	. "github.com/onsi/gomega"
 
-	"github.com/dash0hq/dash0-operator/internal/dash0/util"
+	"github.com/dash0hq/dash0-operator/internal/util"
 )
 
 type ContainerExpectations struct {
